pipelines/components/common/find-first-crash-commit/requirements.txt
GitPython
pyyaml